requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioshutil==1.5
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
fastapi==0.115.6
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.27.0
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.55.7
MarkupSafe==3.0.2
multidict==6.1.0
openai==1.58.1
packaging==24.2
pdf2image==1.17.0
pillow==11.0.0
propcache==0.2.1
py-zerox==0.0.7
pydantic==2.10.4
pydantic_core==2.27.2
PyPDF2==3.0.1
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rpds-py==0.22.3
sniffio==1.3.1
starlette==0.41.3
tiktoken==0.8.0
tokenizers==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
yarl==1.18.3
zipp==3.21.0
